Менеджер для школьной библиотеки2

+++ src/main/java/ru/mera/lib/service/PupilService.java
@@ -3,6 +3,7 @@ package ru.mera.lib.service;
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import ru.mera.lib.entity.Pupil;
+import ru.mera.lib.entity.RecordCard;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -17,6 +18,7 @@ public class PupilService {
     public PupilService(Session session, BufferedReader reader) {
         this.session = session;
         this.reader = reader;
+
     }
 
     private boolean checkCountInput(String input){
@@ -83,21 +85,29 @@ public class PupilService {
 
 
 
-    public void showPupilsList(){
+    List<Pupil> showPupilsList(){
         Query query = session.createQuery("from Pupil");
-        List<Pupil> pupils = query.getResultList();
+        try {
+            List<Pupil> pupils = query.getResultList();
 
-        pupils.forEach(pupil ->
-                System.out.println("Номер: " + pupil.getId()+
-                "        Имя: " + pupil.getName() +
-                "        Класс: " + pupil.getClassNumber() +
-                "        Буква класса: " + pupil.getClassName()));
+            pupils.forEach(pupil -> {
+                System.out.println("Номер: " + pupil.getId() +
+                        "        Имя: " + pupil.getName() +
+                        "        Класс: " + pupil.getClassNumber() +
+                        "        Буква класса: " + pupil.getClassName());
+                System.out.println("------------------------------------------------------");
+            });
+            return pupils;
+
+        } catch (Exception e) {
+            return null;
+        }
+
     }
 
     Pupil showOnePupil(int id){
         Pupil pupil;
         Query query = session.createQuery("from Pupil where id = " + id);
-//        query.setParameter(0, id);
         try {
             pupil = (Pupil) query.getSingleResult();
         } catch (Exception e){
@@ -108,6 +118,28 @@ public class PupilService {
                 "        Класс: " + pupil.getClassNumber() +
                 "        Буква класса: " + pupil.getClassName());
         return pupil;
+    }
+
+    private List<Pupil> findPupilByName(String name){
+        List<Pupil> pupils;
+        Query query = session.createQuery("from Pupil where name like '%" + name + "%'");
+        try {
+            pupils = query.getResultList();
+            return pupils;
+        } catch (Exception e){
+            return null;
+        }
+    }
+
+    private List<Pupil> findPupilByClassNumber(int classNumber){
+        List<Pupil> pupils;
+        Query query = session.createQuery("from Pupil where classNumber = " + classNumber);
+        try {
+            pupils = query.getResultList();
+            return pupils;
+        } catch (Exception e){
+            return null;
+        }
     }
 
     public void addNewPupil() throws IOException{
@@ -164,10 +196,66 @@ public class PupilService {
         int id = inputId("удаления");
         Pupil pupil = showOnePupil(id);
         if (pupil != null){
-            session.getTransaction();
-            session.delete(pupil);
-            session.getTransaction().commit();
-            System.out.println("Ученик успешно удален!");
+            Query query = session.createQuery("from RecordCard where pupilId = " + pupil.getId() +
+                    " and returnDate = null");
+                List<RecordCard> recordCards = query.getResultList();
+
+                if (recordCards.size() > 0) {
+                    System.out.println("Ученик имеет несданные книги!");
+                } else {
+                    session.beginTransaction();
+                    Query query1 = session.createQuery("delete RecordCard where pupilId = " + pupil.getId() +
+                    " and returnDate <> null");
+                    query1.executeUpdate();
+                    session.delete(pupil);
+                    session.getTransaction().commit();
+                    System.out.println("Ученик успешно удален!");
+                }
         } else System.out.println("Ученик не найден!");
+    }
+
+    public void showPupilsMenu() throws IOException {
+        System.out.println("");
+        System.out.println("1. Показать список всех учеников");
+        System.out.println("2. Найти ученика по имени");
+        System.out.println("3. Найти ученика по id");
+        System.out.println("4. Найти ученика по номеру класса");
+        System.out.println("");
+
+        String choice = reader.readLine();
+        switch (choice){
+            case "1":{
+                showPupilsList();
+                break;
+            }
+            case "2":{
+                System.out.println("");
+                List<Pupil> pupils = findPupilByName(inputName());
+                System.out.println("(или часть имени)");
+                if (pupils != null && pupils.size() > 0) {
+                    pupils.forEach(pupil -> showOnePupil(pupil.getId()));
+                } else System.out.println("Ничего не найдено!");
+                break;
+            }
+            case "3":{
+                int pupilId = inputId("просмотра:");
+                Pupil pupil = showOnePupil(pupilId);
+                if (pupil == null) {
+                    System.out.println("Ученик не найден!");
+                }
+                break;
+            }
+            case "4":{
+                System.out.println("");
+                List<Pupil> pupils = findPupilByClassNumber(inputClassNumber());
+                if (pupils != null && pupils.size() > 0){
+                    pupils.forEach(pupil -> showOnePupil(pupil.getId()));
+                } else System.out.println("Ученики не найдены!");
+                break;
+            }
+            default:{
+                System.out.println("Неверный ввод!");
+            }
+        }
     }
 }
